+++ site-cookbooks/mysql/recipes/default.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: mysql
-# Recipe:: install
+# Recipe:: default
 #
 # Copyright 2016, Leonis0813
 #